# swagger/api.yml
openapi: 3.0.3
servers:
  - description: API Localhost Debug
    url: http://localhost:8080
  - description: API Production talfekler.com
    url: https://talfekler.com
info:
  description: Tal Studio API
  version: 1.0.0
  title: Tal Studio
  contact:
    email: berdugogo@gmail.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /api/events/available/:
    get:
      tags:
        - "events"
      summary: "get all future available trainings"
      description: "retrieval of user trainings"
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  future:
                    type: array
                    items:
                      $ref: "#/components/schemas/available"
              examples:
                get_available:
                  $ref: "#/components/examples/get_available"
          description: "available events retrieved successfully"
        "400":
          description: "Cannot retrieve available"
  /api/events/personalcalendar:
    get:
      tags:
        - "events"
      summary: "get all future available trainings"
      description: "retrieval of user trainings"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/personalcalendar"
              examples:
                personalcalendar:
                  $ref: "#/components/examples/personalcalendar"
          description: "available events retrieved successfully"
        "400":
          description: "Cannot retrieve available"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie
  schemas:
...
    training:
      type: object
      properties:
        participant_ids:
          type: array
          items:
            type: string
        participant_names:
          type: array
          items:
            type: string
        title:
          type: string
        _id:
          type: string
        description:
          type: string
        date:
          type: string
          format: DATETIME
    available:
      type: object
      properties:
        _id:
          type: string
        date:
          type: string
          format: DATETIME
        title:
          type: string
        description:
          type: string
...
    personalcalendar:
      type: object
      properties:
        trainings:
          type: array
          items:
            $ref: "#/components/schemas/training"
        available:
          type: object
          properties:
            future:
              type: array
              items:
                $ref: "#/components/schemas/available"
  examples:
...
    get_available:
      value:
        {
          future:
            [
              {
                _id: "6173477c70fc735c5004d009",
                date: "2021-10-29T03:00:00.000Z",
                title: "פנוי",
                description: ""
              },
              {
                _id: "617347b170fc735c5004d01d",
                date: "2021-10-30T04:00:00.000Z",
                title: "פנוי",
                description: ""
              },
            ]
        }
...
    personalcalendar:
      value:
        trainings: 
          [
            {
              participant_ids:
                [
                "61732aaa4675f478b083832a",
                "6173353e8d2a5d277cdb5767"
                ],
              participant_names: 
                [
                "עדן ברדוגו",
                "טל פקלר"
                ],
              title: "אימון",
              _id: "61733690bdbd498a206c9115",
              description": "",
              date: "2021-10-24T04:00:00.000Z"
            }
          ]
        available:
          future:
            [
              {
                _id: "617335858d2a5d277cdb5781",
                date: "2021-10-24T03:00:00.000Z",
                title: "פנוי",
                description: ""
              },
              {
                _id: "617335858d2a5d277cdb5783",
                date: "2021-10-24T05:00:00.000Z",
                title: "פנוי",
                description: ""
              },
            ]
...